Do not handle missing sections
The problem is with a structure like this...

    # {title}

    {foo}

    ## Description

    {description}

...if the input doc does not contain a "Description" section, the description
variable is set to the contents of "foo". OTOH, if we want return an empty array
for description, that messes with trying to get 'foo' because 'foo' has no
obvious corresponding title (just looking for the title directly above is what
causes description to be set to the contents of foo.

Anyway, the solution is that the user should provide good data and validate it
themselves if needed.